se cambia id de cliente a Long y se agrega servicios y mapping de cliente

+++ src/main/java/com/pragma/restaurant/entity/Client.java
@@ -12,7 +12,7 @@ public class Client {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
     @Column(name = "name", nullable = false)
     private String name;
 
@@ -27,18 +27,18 @@ public class Client {
     public Client() {
     }
 
-    public Client(Integer id, String name, boolean orderActive, List<Order> orderUser) {
+    public Client(Long id, String name, boolean orderActive, List<Order> orderUser) {
         this.id = id;
         this.name = name;
         this.orderActive = orderActive;
         this.orderUser = orderUser;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
